🔧 Rename ViTConfig & Updatae line:import ViTConfig
