+++ config/config_factory.py
@@ -1,7 +1,7 @@
 from .config import Config
 from .custom_nn_config import CustomNNConfig
 from .coatnet_config import CoAtNetConfig
-from .vit import ViTConfig
+from .vit_config import ViTConfig
 
 # 모델 이름과 구성 설정 클래스를 매핑하는 디렉토리
 CONFIG_MAP = {
